requirements.txt
pydantic
langchain
openai
python-dotenv
coloredlogs